requirements.txt
blinker==1.5
certifi==2021.10.8
click==7.1.2
colorama==0.4.6
dnspython==1.16.0
Flask==1.1.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pymongo==3.13.0
python-dotenv==0.18.0
Werkzeug==1.0.1
pyTelegramBotAPI==4.15.4
flask-wtf==0.15.0
urllib3<2.0